vanilla/requirements.txt
flask
numba
pillow
numpy
